+++ src/exchange/kraken.rs
@@ -87,7 +87,7 @@ impl Retrieve for Kraken {
         let result = match interval.1 {
             TermUnit::Sec => return Err(anyhow!("OKX does not support candlestick of seconds")),
             TermUnit::Min => interval.0.to_string(),
-            TermUnit::Hour => (interval.0 * 60).to_string(),
+            TermUnit::Hour => (interval.0 as i32 * 60).to_string(),
             TermUnit::Day => (interval.0 as i32 * 60 * 24).to_string(),
             TermUnit::Week => (interval.0 as i32 * 60 * 24 * 7).to_string(),
             TermUnit::Month => return Err(anyhow!("OKX does not support candlestick of months")),
